refactor(caret): extract UserSelectingTracker; CaretModel delegates isUserSelecting
